Refactor Profile event list by moving occurrence info to own component
