+++ src/domain/profile/events/profileEventsList.module.scss
@@ -1,44 +1,10 @@
 @import 'mixins';
 
-$labelIconMargin: 0.5rem;
-$labelColumnMargin: 2rem;
-$labelRowMargin: 1rem;
 $qrMargin: 3rem;
 $eventsListMargin: 4rem;
 
 .eventsList {
   margin-top: $eventsListMargin;
-}
-
-.row {
-  @include respond-above(sm) {
-    flex-direction: row;
-  }
-  display: flex;
-  flex-direction: column;
-}
-
-.label {
-  @include respond-above(sm) {
-    & + & {
-      margin-left: $labelColumnMargin;
-    }
-  }
-  @include respond-below(sm) {
-    & + & {
-      margin-top: $labelRowMargin;
-    }
-  }
-  display: flex;
-  justify-content: normal;
-  flex-direction: row;
-  border: 0;
-  text-align: left;
-  align-items: center;
-}
-
-.labelIcon {
-  margin-right: $labelIconMargin;
 }
 
 .qrWrapper {
